requirements.txt
beautifulsoup4==4.5.1
bs4==0.0.1
click==6.6
cookies==2.2.1
Flask==0.11.1
future==0.15.2
geopy==1.11.0
googlefinance==0.7
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauth2==1.9.0.post1
oauthlib==2.0.0
python-dotenv==0.6.0
python-forecastio==1.3.4
python-twitter==3.1
requests==2.11.1
requests-oauthlib==0.7.0
responses==0.5.1
six==1.10.0
slacker==0.9.25
Werkzeug==0.11.11
yelp==1.0.2